basic_examples/python/requirements.txt
livekit-agents[openai,silero,deepgram,cartesia,turn-detector]~=1.0rc
livekit-plugins-noise-cancellation~=0.2
python-dotenv





